Reduced photodiode data size in memory
- 16 ADC bin number is now directly stored in flash, can be converted to voltage by multiplying by ADC voltage resolution.
-  Modified saveBuffer() to store one huge array of encoded bytes

+++ FSW/src/headers/timing.hpp
@@ -29,8 +29,8 @@ enum Mode { // all payload science modes
 };
 
 /* - - - - - - Declarations - - - - - - */
-void updatePayloadMode(float buffer[BUFFERSIZE], int bufIdx);
-float voltageRunningMean(float buffer[BUFFERSIZE], int bufIdx);
-void checkSweepChange(float buffer[BUFFERSIZE], int bufIdx);
+void updatePayloadMode(uint16_t buffer[BUFFERSIZE], int bufIdx);
+float voltageRunningMean(uint16_t buffer[BUFFERSIZE], int bufIdx);
+void checkSweepChange(uint16_t buffer[BUFFERSIZE], int bufIdx);
 int wrapBufferIdx(int idx);
 #endif
